+++ yampt.core/source/io/file_list.hpp
@@ -11,7 +11,8 @@ enum class file_type_t
 	base_dict,
 	user_dict,
 	yaml_l10n,
-	loc_file
+	loc_file,
+	eet_file
 };
 
 struct file_entry_t